Adding user edit status method

+++ controllers/feed.js
@@ -64,7 +64,6 @@ exports.createPost = (req, res, next) => {
       return user.save();
     })
     .then((result) => {
-      console.log(post);
       res.status(201).json({
         message: "Post created successfully!",
         post: post,
@@ -129,6 +128,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if(post.creator.toString() !==req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
       if (updateImageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -160,12 +164,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      //check logged in user
+      if(post.creator.toString() !==req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
